+++ .devcontainer/devcontainer.json
@@ -34,7 +34,7 @@
 	// "forwardPorts": [],
 
 	// Use 'postCreateCommand' to run commands after the container is created.
-	"postCreateCommand": "pip3 install pip-tools",
+	"postCreateCommand": "pip3 install pip-tools pytest",
 
 	// Configure tool-specific properties.
 	// "customizations": {},
