--- models/video.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-
-class VideoInput(BaseModel):
-    url: str
-
-
-class VideoResult(BaseModel):
-    text: str
