+++ src/summary_eval_supabase.py
@@ -13,7 +13,7 @@ from transformers import logging
 
 from models.summary import SummaryInput, SummaryResults
 from pipelines.summary import SummaryPipeline
-from src.database import get_client
+from connections.database import get_client
 
 nlp = spacy.load("en_core_web_sm", disable=["ner"])
